chore(hygiene): trace-gate the modular DIAG stderr spam + clear the it-binary clippy debt
1. modular/section.rs printed three plain eprintln! DIAG lines (tree /
   code / LfGlobal) on EVERY multi-group lossless encode — visible to
   any CLI user without stderr redirected. Now crate::trace::
   debug_eprintln! like every other diagnostic in the file (timing vars
   underscore-prefixed for the macro-inactive build). CLI stderr is
   clean; bytes untouched (hash-locks 40/40 in the suite run).

2. Cleared all 17 pre-existing clippy lints in the it test binary so
   `clippy -p jxl-encoder --tests --features __internals -- -D
   warnings` is green: mechanical fixes (casts, doc indent, enumerate,
   type alias, 3.14->3.5 test constant) plus two justified file-level
   allows — afv_idct_parity's libjxl reference tables keep verbatim
   digits (approx_constant), and lossy_knobs_wiring constructs
   #[non_exhaustive] EncoderImprovementsCustom via Default-then-assign
   because struct literals cannot compile outside the crate
   (field_reassign_with_default).

it 392/392, workspace clippy clean both configs.

+++ jxl-encoder/tests/it/chroma_subsampling_signal.rs
@@ -96,8 +96,8 @@ fn decode_via_jxl_rs(bytes: &[u8]) {
         }
     }
 
-    assert_eq!(width as usize, W);
-    assert_eq!(height as usize, H);
+    assert_eq!(width, W);
+    assert_eq!(height, H);
 }
 
 fn assert_invalid_config_with_tag(err: &EncodeError, tag: &str) {
